refactor: 영화 예매 화면 개선
