+++ app/src/main/java/woowacourse/movie/presentation/view/reservation/detail/ReservationDetailContract.kt
@@ -45,8 +45,6 @@ interface ReservationDetailContract {
 
         fun notifyNoAvailableDates()
 
-        fun notifyReservationConfirm()
-
-        fun navigateToResult(reservationInfo: ReservationInfoUiModel)
+        fun notifyReservationConfirm(reservationInfo: ReservationInfoUiModel)
     }
 }
